move sliderbars to OsuSettings
