+++ osu.Game/Overlays/Settings/Sections/Gameplay/GeneralSettings.cs
@@ -44,26 +44,8 @@ namespace osu.Game.Overlays.Settings.Sections.Gameplay
                 {
                     LabelText = "Score display mode",
                     Bindable = config.GetBindable<ScoringMode>(OsuSetting.ScoreDisplayMode)
-                },
-                new SettingsSlider<int, TimeSlider>
-                {
-                    LabelText = "Followpoint fadeout time",
-                    TransferValueOnCommit = true,
-                    Bindable = config.GetBindable<int>(OsuSetting.FollowPointAppearTime),
-                    KeyboardStep = 1
-                },
-                new SettingsSlider<int, TimeSlider>
-                {
-                    LabelText = "Followpoint fadeout offset",
-                    TransferValueOnCommit = true,
-                    Bindable = config.GetBindable<int>(OsuSetting.FollowPointDelay),
-                    KeyboardStep = 1
                 }
             };
-        }
-        private class TimeSlider : OsuSliderBar<int>
-        {
-            public override string TooltipText => base.TooltipText + "ms";
         }
     }
 }
